requirements.txt
boto3
python-dotenv
psycopg2-binary
pytest
pylint
logging